requirements.txt
asgiref==3.4.1
cffi==1.15.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.11
django-rest-framework==0.1.0
django-filter==21.1
drf-spectacular==0.22.1
drf-writable-nested==0.6.3
jwt==1.3.1
ldap3==2.9.1
selenium==3.141.0
requests==2.27.1
psycopg2==2.9.6